+++ CHANGELOG.md
@@ -1,7 +1,8 @@
 # 1.x branch
 ## 1.5 branch
 ### 1.5.7
-* some fixes to get it ready for CakePHP 4.3.
+* ready for CakePHP 4.3. The `AssetMiddleware` was moved from
+    `Assets\Routing\Middleware` to `Assets\Middleware`.
 
 ### 1.5.6
 * migration to github actions.